requirements.txt
joblib==1.1.0
numpy==1.20.2
pandas==1.2.4
Pillow==9.0.0
scikit_learn==1.0.2
streamlit==1.3.1
underthesea==1.3.3
